Move NeoForge mod entry point to separate class and package
Rename and move ClientTotemicMod as well
